# docs/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Digital Procurement and PLS Seedbed API
  version: 0.24.0
  summary: Compliance-first procurement evidence API with proof anchoring and PLS seedbed support.
  description: >
    API contract for the supervisor-demo plus selected pilot-hardening baseline.
    The runtime uses opaque bearer sessions. OAuth2/OIDC entries are readiness
    boundaries and do not claim a production identity provider.
...
servers:
  - url: http://localhost:3100/api/v1
    description: Local backend
...
paths:
...
  /company-ledger/exports:
    post:
      tags: [Company Ledger, Productivity]
      operationId: createCompanyLedgerExport
      security:
        - bearerSession: []
      responses:
        "201":
          description: Lightweight company ledger manifest created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyLedgerExportResponse"
  /company-productivity/summary:
    get:
      tags: [Productivity]
      operationId: getCompanyProductivitySummary
      security:
        - bearerSession: []
      responses:
        "200":
          description: Money tracker, pipeline, inbox, scorecards, and checklist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyProductivitySummaryResponse"
  /company-productivity/action-inbox:
    get:
      tags: [Productivity]
      operationId: listActionInbox
      security:
        - bearerSession: []
      responses:
        "200":
          description: Open and completed company tasks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionInboxResponse"
  /company-productivity/action-inbox/{taskId}:
    patch:
      tags: [Productivity]
      operationId: completeActionInboxTask
      security:
        - bearerSession: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Task marked complete.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionInboxItemResponse"
  /company-productivity/saved-views:
    get:
      tags: [Productivity]
      operationId: listSavedWorkspaceViews
      security:
        - bearerSession: []
      responses:
        "200":
          description: Saved views for the current organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedViewsResponse"
    post:
      tags: [Productivity]
      operationId: createSavedWorkspaceView
      security:
        - bearerSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: Proof exceptions }
                filter: { type: string, example: proofStatus in failed,mismatch,unavailable }
      responses:
        "201":
          description: Saved view created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedViewResponse"
...
  /notifications:
    get:
      tags: [Email, Productivity]
      operationId: listNotifications
      security:
        - bearerSession: []
      responses:
        "200":
          description: Safe local notification center records.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationsResponse"
...
components:
  securitySchemes:
    bearerSession:
      type: http
      scheme: bearer
      bearerFormat: opaque-session
      description: Current runtime opaque bearer session token.
    oauth2:
      type: oauth2
      description: OAuth2/OIDC readiness boundary only; no production IdP is configured.
      flows:
        authorizationCode:
          authorizationUrl: https://identity.example.test/oauth/authorize
          tokenUrl: https://identity.example.test/oauth/token
          scopes:
            org:read: Read organization profile and context.
            org:write: Update organization metadata.
            org:users:manage: Manage organization users.
            network:read: Read organization network graph.
            network:write: Create and decide network requests.
            deals:read: Read company ledger projections.
            deals:write: Prepare deal actions.
            productivity:read: Read productivity workspace.
            productivity:write: Update tasks and saved views.
            proof:read: Read proof metadata.
            proof:verify: Verify proof hashes.
            email:read: Read local notification outbox.
            admin:platform: Platform administration.
...
  schemas:
...
    CompanyProductivitySummaryResponse:
      type: object
      properties:
        data:
          type: object
          required: [moneyTracker, pipeline, actionInbox, supplierScorecards, evidenceChecklist]
          properties:
            moneyTracker:
              $ref: "#/components/schemas/MoneyTracker"
            pipeline:
              type: array
              items: { type: object }
            actionInbox:
              type: array
              items:
                $ref: "#/components/schemas/ActionInboxItem"
            supplierScorecards:
              type: array
              items: { type: object }
            evidenceChecklist:
              type: array
              items: { type: object }
    MoneyTracker:
      type: object
      properties:
        currency: { type: string, example: MYR }
        committedPurchaseOrderValue: { type: string, example: "68000.00" }
        expectedDeliveryValue: { type: string }
        escrowAmount: { type: string }
        financingCapitalAmount: { type: string }
        budgetRemaining: { type: string }
        safeAmountLabel: { type: string }
    ActionInboxResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ActionInboxItem"
    ActionInboxItemResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/ActionInboxItem"
    ActionInboxItem:
      type: object
      properties:
        taskId: { type: string }
        title: { type: string }
        description: { type: string }
        priority: { type: string, enum: [low, medium, high] }
        status: { type: string, enum: [open, completed] }
        actionType: { type: string }
    SavedViewsResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/SavedWorkspaceView"
    SavedViewResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/SavedWorkspaceView"
    SavedWorkspaceView:
      type: object
      properties:
        viewId: { type: string }
        organizationId: { type: string }
        name: { type: string }
        filter: { type: string }
        createdByUserId: { type: string }
        createdAt: { type: string, format: date-time }
...
    CompanyLedgerExportResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            exportId: { type: string }
            organizationId: { type: string }
            generatedByUserId: { type: string }
            generatedAt: { type: string, format: date-time }
            itemCount: { type: integer }
            manifestHash: { type: string, example: sha256:abc123 }
            proofScope: { type: string, example: companyLedgerSummary }
            status: { type: string, example: ready }
            safeSummary: { type: string }
    NotificationsResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                type: object
...